+++ include/psen_scan_v2/ros_scanner_node.h
@@ -25,6 +25,8 @@
 #include <gtest/gtest_prod.h>
 
 #include <ros/ros.h>
+#include <std_msgs/UInt8.h>
+#include <sensor_msgs/LaserScan.h>
 
 #include "psen_scan_v2_standalone/scanner_v2.h"
 
@@ -64,7 +66,7 @@ public:
                   const double& x_axis_rotation,
                   const ScannerConfiguration& scanner_config);
 
-  //! @brief Continuously fetches data from the scanner and publishes the data as ROS scanner message.
+  //! @brief Continuously fetches data from the scanner and publishes the data into the ROS network.
   void run();
   //! @brief Terminates the fetching and publishing of scanner data.
   void terminate();
@@ -74,18 +76,21 @@ private:
 
 private:
   ros::NodeHandle nh_;
-  ros::Publisher pub_;
+  ros::Publisher pub_scan_;
+  ros::Publisher pub_zone_;
   std::string tf_prefix_;
   double x_axis_rotation_;
   S scanner_;
   std::atomic_bool terminate_{ false };
 
   friend class RosScannerNodeTests;
-  FRIEND_TEST(RosScannerNodeTests, testScannerInvocation);
-  FRIEND_TEST(RosScannerNodeTests, testScanTopicReceived);
-  FRIEND_TEST(RosScannerNodeTests, testScanBuildFailure);
-  FRIEND_TEST(RosScannerNodeTests, testMissingStopReply);
-  FRIEND_TEST(RosScannerNodeTests, shouldNotInvokeUserCallbackInCaseOfEmptyLaserScan);
+  FRIEND_TEST(RosScannerNodeTests, shouldStartAndStopSuccessfullyIfScannerRespondsToRequests);
+  FRIEND_TEST(RosScannerNodeTests, shouldPublishScansWhenLaserScanCallbackIsInvoked);
+  FRIEND_TEST(RosScannerNodeTests, shouldPublishActiveZonesetWhenLaserScanCallbackIsInvoked);
+  FRIEND_TEST(RosScannerNodeTests, shouldWaitWhenStopRequestResponseIsMissing);
+  FRIEND_TEST(RosScannerNodeTests, shouldReceiveLatchedActiveZonesetMsg);
+  FRIEND_TEST(RosScannerNodeTests, shouldProvideScanTopic);
+  FRIEND_TEST(RosScannerNodeTests, shouldProvideActiveZonesetTopic);
 };
 
 typedef ROSScannerNodeT<> ROSScannerNode;
@@ -101,7 +106,8 @@ ROSScannerNodeT<S>::ROSScannerNodeT(ros::NodeHandle& nh,
   , x_axis_rotation_(x_axis_rotation)
   , scanner_(scanner_config, std::bind(&ROSScannerNodeT<S>::laserScanCallback, this, std::placeholders::_1))
 {
-  pub_ = nh_.advertise<sensor_msgs::LaserScan>(topic, 1);
+  pub_scan_ = nh_.advertise<sensor_msgs::LaserScan>(topic, 1);
+  pub_zone_ = nh_.advertise<std_msgs::UInt8>("active_zoneset", 1, true);
 }
 
 template <typename S>
@@ -117,7 +123,10 @@ void ROSScannerNodeT<S>::laserScanCallback(const LaserScan& scan)
         data_conversion_layer::radianToDegree(laserScanMsg.angle_max),
         data_conversion_layer::radianToDegree(laserScanMsg.angle_increment),
         laserScanMsg.ranges.size());
-    pub_.publish(laserScanMsg);
+    pub_scan_.publish(laserScanMsg);
+    std_msgs::UInt8 active_zoneset;
+    active_zoneset.data = scan.getActiveZoneset();
+    pub_zone_.publish(active_zoneset);
   }
   // LCOV_EXCL_START
   catch (const std::invalid_argument& e)