+++ standalone/include/psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_msg.h
@@ -71,10 +71,12 @@ public:
   Message(const util::TenthOfDegree& from_theta,
           const util::TenthOfDegree& resolution,
           const uint32_t scan_counter,
+          const uint8_t active_zoneset,
           const std::vector<double>& measurements)
     : from_theta_(from_theta)
     , resolution_(resolution)
     , scan_counter_(scan_counter)
+    , active_zoneset_(active_zoneset)
     , measurements_(measurements)
     , diagnostic_data_enabled_(false){
 
@@ -83,12 +85,14 @@ public:
   Message(const util::TenthOfDegree& from_theta,
           const util::TenthOfDegree& resolution,
           const uint32_t scan_counter,
+          const uint8_t active_zoneset,
           const std::vector<double>& measurements,
           const std::vector<double>& intensities,
           const std::vector<data_conversion_layer::monitoring_frame::diagnostic::Message>& diagnostic_messages)
     : from_theta_(from_theta)
     , resolution_(resolution)
     , scan_counter_(scan_counter)
+    , active_zoneset_(active_zoneset)
     , measurements_(measurements)
     , intensities_(intensities)
     , diagnostic_messages_(diagnostic_messages)
@@ -100,6 +104,7 @@ public:
   util::TenthOfDegree fromTheta() const;
   util::TenthOfDegree resolution() const;
   uint32_t scanCounter() const;
+  uint8_t activeZoneset() const;
   const std::vector<double>& measurements() const;
   const std::vector<double>& intensities() const;
   std::vector<data_conversion_layer::monitoring_frame::diagnostic::Message> diagnosticMessages() const;
@@ -109,8 +114,8 @@ private:
   configuration::ScannerId scanner_id_{ configuration::ScannerId::master };
   util::TenthOfDegree from_theta_{ 0 };
   util::TenthOfDegree resolution_{ 0 };
-
   boost::optional<uint32_t> scan_counter_;
+  uint8_t active_zoneset_{ 0 };
   std::vector<double> measurements_;
   std::vector<double> intensities_;
   std::vector<data_conversion_layer::monitoring_frame::diagnostic::Message> diagnostic_messages_;