+++ standalone/main.cpp
@@ -35,6 +35,8 @@ const util::TenthOfDegree ANGLE_END{ data_conversion_layer::degreeToTenthDegree(
  */
 void laserScanCallback(const LaserScan& scan)
 {
+  // Other data fields are listed on
+  // https://docs.ros.org/en/melodic/api/psen_scan_v2/html/classpsen__scan__v2__standalone_1_1LaserScan.html
   PSENSCAN_INFO_THROTTLE(1 /* sec */, "laserScanCallback()", "Ranges {}", util::formatRange(scan.getMeasurements()));
 }