+++ standalone/test/include/psen_scan_v2_standalone/util/integrationtest_helper.h
@@ -72,7 +72,8 @@ static std::vector<double> generateIntensities(const unsigned int& num_elements,
 static data_conversion_layer::monitoring_frame::Message
 createValidMonitoringFrameMsg(const uint32_t scan_counter = 42,
                               const util::TenthOfDegree start_angle = DEFAULT_SCAN_RANGE.getStart(),
-                              const util::TenthOfDegree end_angle = DEFAULT_SCAN_RANGE.getEnd())
+                              const util::TenthOfDegree end_angle = DEFAULT_SCAN_RANGE.getEnd(),
+                              const uint8_t active_zoneset = 1)
 {
   const auto resolution{ util::TenthOfDegree(10) };
 
@@ -90,7 +91,13 @@ createValidMonitoringFrameMsg(const uint32_t scan_counter = 42,
   };
 
   return data_conversion_layer::monitoring_frame::Message(
-      start_angle, resolution, scan_counter, measurements, intensities, diagnostic_messages);
+      start_angle, resolution, scan_counter, active_zoneset, measurements, intensities, diagnostic_messages);
+}
+
+static data_conversion_layer::monitoring_frame::Message
+createValidMonitoringFrameMsgWithZoneset(const uint8_t active_zoneset)
+{
+  return createValidMonitoringFrameMsg(42, DEFAULT_SCAN_RANGE.getStart(), DEFAULT_SCAN_RANGE.getEnd(), active_zoneset);
 }
 
 std::vector<data_conversion_layer::monitoring_frame::Message>