+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Forthcoming
 * Remove dependency on pilz_testutils
 * Always build hardware tests
 * Internal refactorings
+* Add active zoneset to LaserScan
+* Publish active zoneset id to ~/active_zoneset
 
 0.3.2 (2021-09-16)
 ------------------